+++ protobuf/lib/src/protobuf/pb_map.dart
@@ -26,9 +26,7 @@ class PbMap<K, V> extends MapBase<K, V> {
 
   bool _isReadonly = false;
 
-  // The provided [info] will be ignored.
-  PbMap(this.keyFieldType, this.valueFieldType, [BuilderInfo? info])
-      : _wrappedMap = <K, V>{};
+  PbMap(this.keyFieldType, this.valueFieldType) : _wrappedMap = <K, V>{};
 
   PbMap.unmodifiable(PbMap other)
       : keyFieldType = other.keyFieldType,
@@ -44,8 +42,8 @@ class PbMap<K, V> extends MapBase<K, V> {
     if (_isReadonly) {
       throw UnsupportedError('Attempted to change a read-only map field');
     }
-    _checkNotNull(key);
-    _checkNotNull(value);
+    ArgumentError.checkNotNull(key, 'key');
+    ArgumentError.checkNotNull(value, 'value');
     _wrappedMap[key] = value;
   }
 
@@ -111,12 +109,6 @@ class PbMap<K, V> extends MapBase<K, V> {
     var value =
         entryFieldSet._values[1] ?? mapEntryMeta.byIndex[1].makeDefault!();
     _wrappedMap[key] = value;
-  }
-
-  void _checkNotNull(Object? val) {
-    if (val == null) {
-      throw ArgumentError("Can't add a null to a map field");
-    }
   }
 
   PbMap freeze() {